+++ qlib/kernel/fs/inode.rs
@@ -74,7 +74,7 @@ use crate::qlib::kernel::fs::procfs::symlink_proc::SymlinkNode;
 use crate::qlib::kernel::fs::procfs::seqfile::SeqFile;
 use crate::qlib::kernel::fs::tmpfs::tmpfs_file::TmpfsFileInodeOp;
 use crate::qlib::kernel::fs::secretfs::dir_proc::SecDirNode;
-use crate::shielding_layer::*;
+use crate::shield::inode_tracker::*;
 
 pub fn ContextCanAccessFile(task: &Task, inode: &Inode, reqPerms: &PermMask) -> Result<bool> {
     let creds = task.creds.clone();
