+++ qvisor/src/shield/https_attestation_provisioning_cli.rs
@@ -1,0 +1,28 @@
+
+use crate::qlib::common::*;
+use crate::qlib::kernel::task::Task;
+
+
+
+/// The supported TEE types:
+/// - Tdx: TDX TEE.
+/// - Sgx: SGX TEE.
+/// - Sevsnp: SEV-SNP TEE.
+/// - Sample: A dummy TEE that used to test/demo the KBC functionalities.
+#[derive(Debug, Clone)]
+pub enum Tee {
+    Sev,
+    Sgx,
+    Snp,
+    Tdx,
+
+    // This value is only used for testing an attestation server, and should not
+    // be used in an actual attestation scenario.
+    Sample,
+}
+
+
+
+pub fn provisioning_http_client(_task: &Task) -> Result<usize> {
+    Ok(0)
+}
